requirements.txt
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
emoji==1.6.1
filelock==3.4.0
huggingface-hub==0.2.1
idna==3.3
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
pyparsing==3.0.6
python-dotenv==0.19.2
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
tomli==1.2.2
tqdm==4.62.3
tweepy==4.4.0
typing_extensions==4.0.1
urllib3==1.26.7
